Refactor IllegalStudyAreaException into exception package
