+++ src/main/java/exception/IllegalStudyAreaException.java
@@ -1,4 +1,4 @@
-package studyarea;
+package exception;
 
 /**
  * This is an extension of the Exception class that exist only for StudyArea related modules.
