+++ src/test/java/studyarea/StudyAreaListTest.java
@@ -1,5 +1,6 @@
 package studyarea;
 
+import exception.IllegalStudyAreaException;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
